Adding a new utility class: FastLogSum.
This computes sums of logs by calculating the log of the product instead,
while avoiding falling out of the floating point range by exploiting
the fact that on all modern platforms, double precision floats are laid
out in memory with a IEEE754 bit pattern. After each multiplication the
exponent part is collected in an ordinary integer and what remains is a
number between 1.0 and 2.0. After all terms/factors have been added only
a single logarithm is actually computed, instead of a log for all terms.

The speedup that you get with this utility depends on many things, such
as how CPU-intensive the rest of the code (like, the calculation of the
arguments of the log calls, usually PDF values) is, platform (hardware,
OS, compiler), etcetera. Summing the logs of a few hundred numbers that
are computed cheaply can be sped up with a factor of 3.

The class is very simple, defined completely in the header file.
The core operations are only a few lines, but you need to understand
the IEEE754 specification in order to know what is going on... :-)

Unit tests included.



git-svn-id: http://code.icecube.wisc.edu/svn/projects/phys-services/trunk@39100 16731396-06f5-0310-8873-f7f720988828