Do not use fixed hardcoded names for the ROOT histograms that are used in the I3MediumService implementation; but rather define unique names, so that there can be more than one I3MediumService (with different ice models, for example). Very useful for reconstruction studies.
This is a quick fix, which ensures histogram name uniqueness by using the
"instantation number".

Longer term fixes would demand the user to provide a basename for the
histograms (e.g. the ice model name). Even better it would be to remove the
ROOT histograms altogether and use any other smart array type, which does not
suffer from the ROOT dictionary bug.



git-svn-id: http://code.icecube.wisc.edu/svn/projects/phys-services/trunk@80777 16731396-06f5-0310-8873-f7f720988828