Bring forward I3CutsValues fixes from R98794
Here's the summary.

Attempt to fix breakage caused by r95024-95026:
Proceed to use a pointer only when it is non-NULL, rather than the reverse.
Fix serialization for I3CutsValues; however, previous versions of I3CutsValues
did not store a version number. As such, old versions of the code have no
safety check to ensure that the version being deserialized is one they
understand, so when new data is read by the old code silent corruption will
occur.

I also included the latest changes to the GCDAuditor for overall happy
feelings (or paranoia) when checking GCD files.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/phys-services/releases/V13-01-00@98851 16731396-06f5-0310-8873-f7f720988828