A new parameter was added to enable only checking events within the good run times.
If you enable it, it tries to read the good run times from the D frame. The good run times are added within the L2 offline processing.

This patch is required since the following can happen: At the end of a run some DOMs dropped. Therefore, the good run end time has been set right before when the DOMs dropped. If the good end time is before the file that is tested, it won't find any DOM hits despite those DOMs aren't in the BadDomList -> error. It's a rare case but it can also improve computing time since it doesn't require to check events after the good end time.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/phys-services/trunk@147413 16731396-06f5-0310-8873-f7f720988828